load/requirements.txt
locust>=2.29.0
python-dotenv>=1.0.0
